mcp-server-6/pyproject.toml
[project]
name = "github-releases-tags"
version = "0.1.0"
description = "MCP сервер для анализа активности репозиториев GitHub"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "aiolimiter>=1.1.0",
    "prometheus-client>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "respx>=0.20.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

